+++ python/py_model_couplers.cc
@@ -58,12 +58,27 @@ namespace {
                 const std::string & element_group) {
                self.applyBC(func, element_group);
              })
-
+        .def("getSolidMechanicsModel",
+             &CouplerSolidContact_::getSolidMechanicsModel,
+             py::return_value_policy::reference)
+        .def("getContactMechanicsModel",
+             &CouplerSolidContact_::getContactMechanicsModel,
+             py::return_value_policy::reference)
+        .def("getDisplacement", &CouplerSolidContact_::getDisplacement,
+             py::return_value_policy::reference)
+        .def("getVelocity", &CouplerSolidContact_::getVelocity,
+             py::return_value_policy::reference)
+        .def("getAcceleration", &CouplerSolidContact_::getAcceleration,
+             py::return_value_policy::reference)
+        .def("getMass", &CouplerSolidContact_::getMass,
+             py::return_value_policy::reference)
+        .def("getContactForce", &CouplerSolidContact_::getContactForce,
+             py::return_value_policy::reference)
+        .def("getContactDetector", &CouplerSolidContact_::getContactDetector,
+             py::return_value_policy::reference)
         .def("setTimeStep", &CouplerSolidContact_::setTimeStep,
-             py::arg("time_step"), py::arg("solver_id") = "");
-    // .def("getContactMechanicsModel",
-    //      &CouplerSolidContact_::getContactMechanicsModel,
-    //      py::return_value_policy::reference);
+             py::arg("time_step"), py::arg("solver_id") = "")
+        .def("getStableTimeStep", &CouplerSolidContact_::getStableTimeStep)
   }
 } // namespace